requirements.txt
anyio==3.7.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cloudscraper==1.2.71
colorama==0.4.6
emoji==2.6.0
exceptiongroup==1.1.2
gingerit==0.9.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
MarkupSafe==2.1.3
nltk==3.8.1
numpy==1.25.1
pyparsing==3.1.0
python-dateutil==2.8.2
python-telegram-bot==20.4
pytz==2023.3
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.0
textblob==0.17.1
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.6
zipp==3.16.2